Basic implementation of PSKC class
This class is used for handling PSKC files. It will parse the file and
store relevant properties for easy access. The Key class corresponds to
a single key defined in the PSKC file.

This is a very minimal implementation that only provides some meta-data
from the file and keys (work in progress).