Provide pskc.key docstrings
This documents most of the information that is available per key and
adds a few other minor cosmetic changes.

This also re-organises the key properties to be in a slightly more
logical order and renames the userid key property to key_userid to more
clearly distinguish it from device_userid.