Use original examples from old profiles Internet Draft
This updates the tests to use the original examples from
draft-hoyer-keyprov-pskc-algorithm-profiles-01 instead of modifying them
to fit the RFC 6030 schema (but does include some minor changes to make
them valid XML).

This adds a few additions to the parser to handle legacy challenge and
resposne encoding and a few key policy properties.

This also includes a fix for 0b757ec in the handling of the
<ChallengeFormat> element under a <Usage> element.